refactor: move single-page widgets into pages
